requirements.txt
fastapi
uvicorn
python-dotenv
mistralai==0.4.2
pydantic
sqlalchemy
jinja2
reportlab
python-multipart
# for email
secure-smtplib
